pyproject.toml
[tool.poetry]
name = "gpt-excel-translator"
version = "0.1.0"
description = ""
authors = ["yunyii <yunyi.zao@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt_excel_translator"}]

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.3"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
fire = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
